+++ src/api/projects/merge_requests.rs
@@ -8,15 +8,19 @@
 //!
 //! These endpoints are used for querying projects merge requests.
 
+mod approve;
 pub mod awards;
 mod create;
 pub mod discussions;
 mod edit;
 mod issues_closed_by;
+mod merge;
 mod merge_request;
 mod merge_requests;
 pub mod notes;
+mod rebase;
 mod resource_label_events;
+mod unapprove;
 
 pub use self::create::CreateMergeRequest;
 pub use self::create::CreateMergeRequestBuilder;
@@ -30,6 +34,17 @@ pub use self::issues_closed_by::IssuesClosedByBuilder;
 
 pub use self::merge_request::MergeRequest;
 pub use self::merge_request::MergeRequestBuilder;
+
+pub use self::rebase::RebaseMergeRequest;
+pub use self::rebase::RebaseMergeRequestBuilder;
+
+pub use self::merge::MergeMergeRequest;
+pub use self::merge::MergeMergeRequestBuilder;
+
+pub use self::approve::ApproveMergeRequest;
+pub use self::approve::ApproveMergeRequestBuilder;
+pub use self::unapprove::UnapproveMergeRequest;
+pub use self::unapprove::UnapproveMergeRequestBuilder;
 
 pub use self::merge_requests::MergeRequestOrderBy;
 pub use self::merge_requests::MergeRequestScope;